requirements.txt
fastapi
uvicorn
pydantic
PyYAML
pandas
scikit-learn
xgboost
joblib
matplotlib
seaborn
flask
scikit-optimize